requirements.txt
# Anti-Deepfake-Box Dependencies
# Install with: pip install -r requirements.txt
# Third-party repos are managed as git submodules (see third_party/)

# ------------------------------------------------------------------ #
#  Core deep learning                                                  #
# ------------------------------------------------------------------ #
torch>=2.1.0
torchvision>=0.16.0

# ------------------------------------------------------------------ #
#  Face detection & alignment (SSOT pipeline)                         #
# ------------------------------------------------------------------ #
insightface>=0.7.3
onnxruntime-gpu>=1.16.0   # for insightface GPU inference; use onnxruntime for CPU

# ------------------------------------------------------------------ #
#  Video & audio I/O                                                   #
# ------------------------------------------------------------------ #
decord>=0.6.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
imageio>=2.31.0

# ------------------------------------------------------------------ #
#  Audio processing (SyncNet / LatentSync)                            #
# ------------------------------------------------------------------ #
openai-whisper>=20231117
librosa>=0.10.0

# ------------------------------------------------------------------ #
#  Signal processing (rPPG SNR)                                       #
# ------------------------------------------------------------------ #
scipy>=1.11.0
numpy>=1.24.0

# ------------------------------------------------------------------ #
#  Evaluation & ML utilities                                           #
# ------------------------------------------------------------------ #
scikit-learn>=1.3.0
Pillow>=10.0.0

# ------------------------------------------------------------------ #
#  Configuration & CLI                                                 #
# ------------------------------------------------------------------ #
PyYAML>=6.0
tqdm>=4.65.0

# ------------------------------------------------------------------ #
#  Optional: rPPG-Toolbox models (install from third_party)           #
# ------------------------------------------------------------------ #
# pip install -e third_party/rppg-toolbox

# ------------------------------------------------------------------ #
#  Optional: Diffusion model datasets (ForgeryNet, GenFace)           #
# ------------------------------------------------------------------ #
# Follow dataset-specific instructions for access

# ------------------------------------------------------------------ #
#  Development                                                         #
# ------------------------------------------------------------------ #
pytest>=7.4.0
